hpo_ingest/hpo-ingest-service/requirements.txt
Flask
google-cloud-storage
google-cloud-bigquery
pandas
pyarrow
